// src/cast.types.d.ts
export interface CastMediaStatus {
    activeTrackIds: number[] | null;
    playerState: PlayerState;
    idleReason: number;
    isMuted: boolean;
    playbackRate: number;
    streamPosition: number;
    volume: number;
    currentItemID: any;
    loadingItemID: any;
    preloadedItemID: any;
    queueData: QueueData;
    queueItemCount: number;
    queueHasPreviousItem?: any;
    queueHasNextItem?: any;
}
// ...
export declare enum PlayerState {
    UNKNOWN = "UNKNOWN",
    IDLE = "IDLE",
    PLAYING = "PLAYING",
    PAUSED = "PAUSED",
    BUFFERING = "BUFFERING",
    LOADING = "LOADING"
}
export declare enum ResumeState {
    UNCHANGED = "UNCHANGED",
    PLAY = "PLAY",
    PAUSE = "PAUSE"
}
export declare enum CastEvent {
    onProviderAdded = "onProviderAdded",
    onProviderChanged = "onProviderChanged",
    onProviderRemoved = "onProviderRemoved",
    onRouteAdded = "onRouteAdded",
    onRoutePresentationDisplayChanged = "onRoutePresentationDisplayChanged",
    onRouteRemoved = "onRouteRemoved",
    onRouteSelected = "onRouteSelected",
    onRouteUnselected = "onRouteUnselected",
    onSessionEnded = "onSessionEnded",
    onSessionEnding = "onSessionEnding",
    onSessionResumed = "onSessionResumed",
    onSessionResuming = "onSessionResuming",
    onSessionStarted = "onSessionStarted",
    onSessionStartFailed = "onSessionStartFailed",
    onSessionStarting = "onSessionStarting",
    onSessionSuspended = "onSessionSuspended",
    onDeviceVolumeChanged = "onDeviceVolumeChanged",
    onDeviceChanged = "onDeviceChanged",
    onMediaStatusChanged = "onMediaStatusChanged",
    mediaQueueWillChange = "mediaQueueWillChange",
    itemsReloaded = "itemsReloaded",
    itemsInsertedInRange = "itemsInsertedInRange",
    itemsUpdatedAtIndexes = "itemsUpdatedAtIndexes",
    itemsRemovedAtIndexes = "itemsRemovedAtIndexes",
    mediaQueueChanged = "mediaQueueChanged"
}
export declare enum MetadataType {
    MOVIE = "MOVIE",
    TV_SHOW = "TV_SHOW",
    MUSIC_TRACK = "MUSIC_TRACK",
    PHOTO = "PHOTO",
    USER = "USER",
    AUDIO_BOOK_CHAPTER = "AUDIO_BOOK_CHAPTER",
    GENERIC = "GENERIC"
}
export declare enum StreamType {
    BUFFERED = "BUFFERED",
    LIVE = "LIVE",
    UNKNOWN = "UNKNOWN",
    NONE = "NONE"
}
export declare enum QueueType {
    GENERIC = "GENERIC",
    ALBUM = "ALBUM",
    PLAYLIST = "PLAYLIST",
    AUDIO_BOOK = "AUDIO_BOOK",
    RADIO_STATION = "RADIO_STATION",
    PODCAST_SERIES = "PODCAST_SERIES",
    TV_SERIES = "TV_SERIES",
    VIDEO_PLAYLIST = "VIDEO_PLAYLIST",
    LIVE_TV = "LIVE_TV",
    MOVIE = "MOVIE"
}
export declare enum RepeatMode {
    OFF = "OFF",
    SINGLE = "SINGLE",
    ALL = "ALL",
    ALL_AND_SHUFFLE = "ALL_AND_SHUFFLE"
}
export interface QueueData {
    name?: string | null;
    queueID?: string | number;
    queueType?: QueueType;
    repeatMode?: RepeatMode;
    startIndex?: number;
    startTime?: number;
}